Update response_time and turnaround_time metrics to be based off ts not first_ts
The problem with using first_ts when calculating response time is that it gives a distorted picture of how fast queues are. e.g. if an item gets put in the error queue then retried, we get this weird anomalous response time spike.

Using ts is more appropriate as it gives a true indication of how long it takes the queue to pick up an item immediately after it was pushed.